% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mthemer.R
\name{mthemer}
\alias{mthemer}
\title{mtheme styled ggplots}
\usage{
mthemer()
}
\value{
a ggplot2 theme
}
\description{
Style ggplot2 plots to match mtheme
}
\details{
mtheme is an excellent lightweight beamer theme for latex
slides. This package complements it with a matchin ggplot2 theme.
}
\examples{
library(ggplot2)
ggplot(iris) +
geom_point(aes(x = Sepal.Length, y = Sepal.Width)) +
mthemer()
}

