% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mthemer.R
\name{scale_colour_m1}
\alias{scale_color_m1}
\alias{scale_colour_m1}
\title{mthemer discrete colors}
\usage{
scale_colour_m1()
}
\value{
a scale_color_manual
}
\description{
An alternative discrete color scale matching mthemer
}
\details{
An alternative discrete color scale matching the mthemer ggplot2
theme that goes nicely with the mtheme beamer theme.
}

